:recycle: Refactor(admin): 관리자 확인 로직 모듈화
Related to: #142
